config refactor
