+++ examples/smart_short_long_agents_notebook.py
@@ -23,17 +23,17 @@ import elfpy.markets.hyperdrive.hyperdrive_actions as hyperdrive_actions
 import elfpy.utils.outputs as output_utils
 import elfpy.utils.post_processing as post_processing
 import elfpy.utils.sim_utils as sim_utils
-import elfpy.simulators as simulators
 import elfpy.types as types
 
 from elfpy import WEI
 from elfpy.math import FixedPoint, FixedPointMath
+from elfpy.simulators.config import Config
 
 # %% [markdown]
 # ### Setup experiment parameters
 
 # %%
-config = simulators.Config()
+config = Config()
 
 # General config parameters
 config.title = "Hyperdrive smart agent demo"
@@ -301,7 +301,7 @@ class LPAgent(elf_agent.Agent):
 
 # %%
 def get_example_agents(
-    rng: NumpyGenerator, experiment_config: simulators.Config, existing_agents: int = 0
+    rng: NumpyGenerator, experiment_config: Config, existing_agents: int = 0
 ) -> list[elf_agent.Agent]:
     """Instantiate a set of custom agents"""
     agents = []
