+++ tests/utils/test_sim_utils.py
@@ -7,11 +7,11 @@ import unittest
 import elfpy.markets.hyperdrive.hyperdrive_actions as hyperdrive_actions
 import elfpy.markets.hyperdrive.hyperdrive_market as hyperdrive_market
 import elfpy.pricing_models.hyperdrive as hyperdrive_pm
-import elfpy.simulators.simulators as simulators
 import elfpy.time as time
 import elfpy.utils.outputs as output_utils
 import elfpy.utils.sim_utils as sim_utils
 from elfpy.math import FixedPoint
+from elfpy.simulators.config import Config
 
 # pylint: disable=too-many-locals
 
@@ -28,7 +28,7 @@ class SimUtilsTest(unittest.TestCase):
             for target_fixed_apr in (0.01, 0.03, 0.05, 0.10, 0.25, 0.5, 1.0, 1.1):
                 for num_position_days in [90, 365]:
                     for pricing_model_name in ["Hyperdrive"]:
-                        config = simulators.Config()
+                        config = Config()
                         config.pricing_model_name = pricing_model_name
                         config.target_liquidity = target_liquidity
                         config.curve_fee_multiple = 0.1
